requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.0
numpy>=1.24.0
opencv-contrib-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pandas>=2.0.0
Flask>=3.0.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
